requirements.txt
numpy>=1.26.0
faiss-cpu>=1.8.0
pytest>=8.2.0
